[refactor] #10 코드리뷰
- 코드 리뷰 코멘트 반영
