+++ src/types/navigations.ts
